Make alltests make target run properly.
This is a nasty hack.  The deal program relies on the current working
directory and we added code to change that to /usr/share/deal to get
everything to work properly afater installing the program rpm.  This
test relies on the current working directory, which is wrong, since the
test directory is not installed with the program.

The change here sets the absolute path to my home.  This will fail for
anyone else running the alltests target without patching it.