Do not chdir() on start-up.
We now look at the value passed in from the -i option on the command
line.  If it sytarts with one of our special directory names, we prepend
the installation directory before calling the interpreter to source it.
This means we don't care what the current directory is any more.

To support this, we had to use the absolute path in the start-up
deal.tcl file.  This should really be changed to check for the existence
of the installation directory before using that path.